correct bugs6

+++ src/main/java/ru/practicum/shareit/item/ItemRepository.java
@@ -28,7 +28,7 @@ public class ItemRepository {
         return item;
     }
 
-    public Item findItemById(Long itemId) {
+    public Item findItemById(Integer itemId) {
         List<List<Item>> list = new ArrayList<>(items.values());
         for (int i = 0; i < list.size(); i++) {
             List<Item> listItem = list.get(i);
